# docs/openapi/swagger.yaml
definitions:
...
  api_rest_auth.AuthResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/codeberg_org_algopatterns_server_algopatterns_users.User'
    type: object
...
  api_rest_auth.UpdateProfileRequest:
    properties:
      avatar_url:
        maxLength: 500
        type: string
      name:
        maxLength: 100
        type: string
    required:
    - name
    type: object
  api_rest_auth.UserResponse:
    properties:
      user:
        $ref: '#/definitions/codeberg_org_algopatterns_server_algopatterns_users.User'
    type: object
...
  codeberg_org_algopatterns_server_algopatterns_users.User:
    properties:
      ai_features_enabled:
        type: boolean
      avatar_url:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      provider:
        type: string
      training_consent:
        type: boolean
      updated_at:
        type: string
    type: object
...
  codeberg_org_algopatterns_server_internal_errors.ErrorResponse:
    properties:
      details:
        description: optional details (sanitized in production)
        type: string
      error:
        description: error code (e.g., "unauthorized", "not_found")
        type: string
      message:
        description: user-friendly message
        type: string
      request_id:
        description: echoed from request for correlation
        type: string
    type: object
host: algopatterns.dev
info:
  contact:
    name: API Support
    url: https://codeberg.org/algopatterns/server
  description: |-
    AI-powered Strudel code generation and collaborative live coding platform

    Features:
    - AI-powered Strudel code generation from natural language
    - Real-time collaborative editing via WebSockets
    - OAuth authentication (Google, GitHub)
    - Anonymous session support
    - Save and share Strudel patterns
  license:
    name: GPL-3.0
    url: https://www.gnu.org/licenses/gpl-3.0.html
  title: Algopatterns API
  version: "1.0"
paths:
...
  /api/v1/auth/{provider}:
    get:
      description: Begin OAuth authentication flow with specified provider (google,
        github, apple)
      parameters:
      - description: OAuth provider
        enum:
        - google
        - github
        - apple
        in: path
        name: provider
        required: true
        type: string
      - description: URL to redirect to after authentication
        in: query
        name: redirect_url
        type: string
      responses:
        "302":
          description: Redirect to OAuth provider
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/codeberg_org_algopatterns_server_internal_errors.ErrorResponse'
      summary: Start OAuth authentication
      tags:
      - auth
  /api/v1/auth/{provider}/callback:
    get:
      description: OAuth provider callback. Redirects to original URL with token,
        or returns JSON if no redirect URL
      parameters:
      - description: OAuth provider
        enum:
        - google
        - github
        - apple
        in: path
        name: provider
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_rest_auth.AuthResponse'
        "302":
          description: Redirect to original URL with token
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/codeberg_org_algopatterns_server_internal_errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/codeberg_org_algopatterns_server_internal_errors.ErrorResponse'
      summary: OAuth callback
      tags:
      - auth
...
  /api/v1/auth/me:
    get:
      description: Get authenticated user's profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_rest_auth.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/codeberg_org_algopatterns_server_internal_errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/codeberg_org_algopatterns_server_internal_errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - auth
    put:
      consumes:
      - application/json
      description: Update authenticated user's name and avatar
      parameters:
      - description: Profile update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api_rest_auth.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_rest_auth.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/codeberg_org_algopatterns_server_internal_errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/codeberg_org_algopatterns_server_internal_errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/codeberg_org_algopatterns_server_internal_errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: 'JWT token for authenticated requests. Format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"